autoscaler/requirements.txt
requests==2.31.0
prometheus-client==0.19.0
docker==6.1.3
python-dotenv==1.0.1
schedule==1.2.1
pyyaml==6.0.1 